+++ src/pages/main/posts/components/Post.tsx
